+++ src/components/speechDataTable.vue
@@ -88,39 +88,22 @@ const props = defineProps({
   type: String,
 });
 
-const replaceNewLine = (str) => {
-  return str.replace(/\n/g, "<br><br>");
-};
 
-const replaceWordWithBoldTags = (str, word) => {
-  // add bold tags to word hits. Only matches whole words
-  // which might not be ideal for lemmatized matches
-  const regex = new RegExp(`(?<!\\p{L})${word}(?!\\p{L})`, "giu");
-  const res = str.replace(regex, `<b>${word}</b>`);
-  return res;
-};
+
+
 
 const displayedData = ref({});
 const rows = ref([]);
 const columns = ref([]);
 
-const originalSpeechText = ref("");
+
 
 const loading = ref(false);
 
 const expandRow = async (props) => {
   props.expand = !props.expand;
-
-  if (props.expand) {
-    const speechData = await speechStore.getSpeech(props.row.id);
-    originalSpeechText.value = speechData.speech_text;
-    const speech = replaceNewLine(originalSpeechText.value);
-    const highlightedSpeech = replaceWordWithBoldTags(speech, props.row.hit);
-    speakerNote.value = speechData.speaker_note;
-    speechText.value = highlightedSpeech;
-  }
-
 };
+
 watchEffect(async () => {
   if (metaStore.submitEvent || props.dataLoaded) {
     loading.value = true;
@@ -210,12 +193,7 @@ watchEffect(async () => {
 });
 const pagination = ref({});
 
-function downloadCurrentSpeech(currentProps) {
-  downloadStore.downloadCurrentSpeechText(
-    originalSpeechText.value,
-    currentProps
-  );
-}
+
 
 function downloadSpeeches() {
   // Accessing the current page from the pagination object
